Revert "feature(ui-refactor-pbehavior-form): Resolve merge conflict + Fix default values"
This reverts commit baa22d710cf30667af0306764753c3df4a543282.

+++ sources/webcore/src/canopsis-next/src/components/other/view/buttons/delete-tab-btn.vue
@@ -12,7 +12,12 @@
 <script>
 import { MODALS } from '@/constants';
 
+import modalMixin from '@/mixins/modal';
+
 export default {
+  mixins: [
+    modalMixin,
+  ],
   props: {
     tab: {
       type: Object,
@@ -29,7 +34,7 @@ export default {
   },
   methods: {
     showDeleteTabModal(tab) {
-      this.$modals.show({
+      this.showModal({
         name: MODALS.confirmation,
         config: {
           action: () => this.deleteTab(tab._id),
